change: outline some implicitly-inlined functions
It's good practice to avoid inlining and depend on TU symbols to exist,
therefore, this patch moves any inlined class definitions to their
own sources.
