+++ src/http/server.hpp
@@ -38,6 +38,7 @@ using unix = boost::asio::local::stream_protocol;
 namespace webvirt::http
 {
 
+/* An HTTP server */
 class server
 {
 private:
@@ -62,58 +63,40 @@ private:
     thread::worker_pool pool_;
 
 public:
-    server(std::filesystem::path socket_path)
-        : socket_path_(std::move(socket_path))
-        , io_(new io_context())
-        , acceptor_(*io_, socket_path_.string())
-        , socket_(*io_)
-        , pool_(*io_)
-    {
-    }
+    /** Construct a server without a webvirt::http::io_context
+     *
+     * @param socket_path Path to unix socket
+     **/
+    server(std::filesystem::path socket_path);
 
-    server(io_context &io, std::filesystem::path socket_path)
-        : socket_path_(std::move(socket_path))
-        , io_owned_(false)
-        , io_(&io)
-        , acceptor_(*io_, socket_path_.string())
-        , socket_(*io_)
-        , pool_(*io_)
-    {
-    }
+    /** Construct a server with an external webvirt::http::io_context
+     *
+     * @param io External webvirt::http::io_context
+     * @param socket_path Path to unix socket
+     **/
+    server(io_context &io, std::filesystem::path socket_path);
 
-    ~server()
-    {
-        pool_.join();
+    /** Destruct a server */
+    ~server();
 
-        if (io_owned_)
-            delete io_;
-    }
+    /** Set connection timeout
+     *
+     * @param ms Timeout in milliseconds
+     * @returns Reference to this
+     **/
+    server &timeout(std::chrono::milliseconds ms);
 
-    server &timeout(std::chrono::milliseconds ms)
-    {
-        timeout_ = ms;
-        return *this;
-    }
+    /** Return connection timeout
+     *
+     * Returns connection timeout
+     **/
+    std::chrono::milliseconds timeout() const;
 
-    std::chrono::milliseconds timeout() const
-    {
-        return timeout_;
-    }
-
-    std::size_t run()
-    {
-        logger::info(fmt::format("Listening on '{}'", socket_path_.c_str()));
-
-        // Begin async calls
-        async_accept();
-
-        // Start up internal worker_pool
-        const auto threads = config::ref().get<unsigned>("threads");
-        pool_.start(threads);
-
-        // Run the io_context on the main thread
-        return io_->run();
-    }
+    /** Run the server's io_context
+     *
+     * @returns Number of handlers processed
+     **/
+    std::size_t run();
 
     handler_setter(on_accept, on_accept_);
     handler_setter(on_request, on_request_);
@@ -124,29 +107,7 @@ public:
     handler_setter(on_close, on_close_);
 
 private:
-    void async_accept()
-    {
-        acceptor_.async_accept(socket_, [this](boost::beast::error_code) {
-            // Here, we don't handle beast::error_code; we let that
-            // job fall through to the connection we make. If there's
-            // a problem with the socket, operations will fails within
-            // the connection immediately, which calls on_error_.
-            auto conn = std::make_shared<connection>(
-                *io_, std::move(socket_), timeout_);
-
-            on_accept_(conn);
-            conn->on_accept(on_accept_);
-            conn->on_request(on_request_);
-            conn->on_websock_accept(on_websock_accept_);
-            conn->on_handshake(on_handshake_);
-            conn->on_websock_read(on_websock_read_);
-            conn->on_error(on_error_);
-            conn->on_close(on_close_);
-            conn->start();
-
-            async_accept();
-        });
-    }
+    void async_accept();
 };
 
 }; // namespace webvirt::http
