+++ src/meson.build
@@ -30,9 +30,13 @@ sources = [
   'virt/connection.cpp',
   'virt/util.cpp',
   'ws/pool.cpp',
+  'ws/client.cpp',
   'ws/connection.cpp',
   'http/router.cpp',
   'http/middleware.cpp',
+  'http/server.cpp',
+  'http/client.cpp',
+  'http/connection.cpp',
   'http/io_context.cpp',
   'http/util.cpp',
   'thread/worker_pool.cpp',
